dedicated unit test suite for basic integrity check on default ai providers
